Align output position with the main base of the kmer
Add the kmer length and central position
(i.e. the base from the kmer that contributes
the most to the kmer current measurement) as
properties of the kit. Those are then used
by the different preprocessors so that the
position in nanocompore's output consistently
has the meaning of "position of the kmer's
central base".

Some bugs in uncalled4 parsing are fixed, e.g.
only the outer edges of the first and last
segments need to be trimmed, not the internal
ones.